excel appending function : started
